Fix loading snappy when not in system library path
Loading the snappy library via System.loadLibrary() succeeds if the library
is in the Hadoop lib/native directory (and thus in the java.library.path).
However, the dlopen() call in the C code fails if the library isn't in the
system library path.  The Java code ignores this error and goes on to use
the library, which then crashes the JVM.

This patch does three things:

1) Use RTLD_DEFAULT instead of dlopen(), as the snappy library is already
   loaded via System.loadLibrary().
2) Initialize the native code in LoadSnappy, allowing the entire loading
   process can be guarded by isLoaded().
3) Add isLoaded() guards to the SnappyCompressor and SnappyDecompressor
   constructors, preventing the JVM from crashing if they are used directly.